pyproject.toml
[tool.poetry]
name = "twitch-bot"
version = "0.1.0"
description = ""
authors = ["Srpboyz <srpboyz69@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PyQt6 = "^6.5.2"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
